api/requirements.txt
Flask
Flask-PyMongo
python-dotenv 
flasgger 